% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Interactively format data}
\usage{
tidyData(data)
}
\arguments{
\item{data}{(Optional) Data frame to be manipulated. If not provided the data from 
all data.frame objects in the current environment will be available for selection in the gadget.}
}
\description{
By selecting a data set the user can set all options for working with the \code{gather}
function of the \code{tidyr} package. The user is able to visualise the impact of the
options set and on return the shiny gadget will return the code required to recreate.
}
\author{
Aimee Gott
}

